+++ src/features/mypage/api/db.json
@@ -5,8 +5,9 @@
       "name": "집",
       "phone": "010-1234-1233",
       "address": "서울특별시 강남구 테헤란로 1234",
-      "detail": "101호",
-      "isDefault": true
+      "detail": "105호",
+      "isDefault": true,
+      "deliveryRequest": "부재시 문앞 연락주세요"
     },
     {
       "id": "addr-2",
@@ -14,7 +15,8 @@
       "phone": "010-1234-1234",
       "address": "부산광역시 해운대구 센텀중앙로 55",
       "detail": "302동 502호",
-      "isDefault": false
+      "isDefault": false,
+      "deliveryRequest": ""
     },
     {
       "id": "addr-3",
@@ -22,7 +24,8 @@
       "phone": "010-1234-1111",
       "address": "대구광역시 수성구 동대구로 100",
       "detail": "1203호",
-      "isDefault": false
+      "isDefault": false,
+      "deliveryRequest": ""
     },
     {
       "id": "addr-4",
@@ -30,7 +33,8 @@
       "phone": "010-4444-5555",
       "address": "인천광역시 연수구 송도과학로 77",
       "detail": "힐스테이트 302동 802호",
-      "isDefault": false
+      "isDefault": false,
+      "deliveryRequest": ""
     },
     {
       "id": "addr-5",
@@ -38,7 +42,8 @@
       "phone": "010-7777-9999",
       "address": "광주광역시 서구 무진대로 45",
       "detail": "501호",
-      "isDefault": false
+      "isDefault": false,
+      "deliveryRequest": ""
     }
   ]
 }
